+++ python/GafferSceneUI/SceneInspector.py
@@ -300,9 +300,10 @@ class SceneInspector( GafferSceneUI.SceneEditor ) :
 
 GafferUI.Editor.registerType( "SceneInspector", SceneInspector )
 
-# InspectorTree isn't public API. Expose the `registerInspectors()` function on SceneInspector
-# itself to make it available to extension authors.
+# InspectorTree isn't public API. Expose the `registerInspectors()` function and the `Inspection` class
+# on SceneInspector itself to make them available to extension authors.
 SceneInspector.registerInspectors = _GafferSceneUI._SceneInspector.InspectorTree.registerInspectors
+SceneInspector.Inspection = _GafferSceneUI._SceneInspector.InspectorTree.Inspection
 
 ##########################################################################
 # Settings metadata